Implemented feature FavoriteDish. But if dish already is in favorites the add button doesn	 disappear. There was also changed some URLs and removed function (/deskByInstitution/institutionId/) from InstitutionController because existed (/desk/Id) in DeskController with similiar functional

+++ cafe/src/main/java/com/master_diploma/cafe/repositories/DishRepository.java
@@ -9,5 +9,5 @@ import java.util.Set;
 @Repository
 public interface DishRepository extends CrudRepository<Dish, Long> {
 
-    Set<Dish> findByInstitutionId(int institutionId);
+    Set<Dish> findByInstitutionId(long institutionId);
 }
